🎨 💄 Convert error page to error component
Make adjustments to the mobile layout
