+++ apps/client/src/Components/Error/Image/ErrorCircle.png
