+++ apps/client/src/Components/Error/Image/index.tsx
