+++ apps/client/src/Components/Error/Image/styles.ts
