+++ apps/client/src/Components/Error/Web/style.ts
